// src/app/interfaces/teachersInterface.ts
export interface Teachers {
    firstName:      string;
    lastName:       string;
    fatherName:     string;
    dateOfBirth:    Date;
    subjects:       string[];
    experience:     number;
    phoneNumber:    string;
    address:        string;
    gender:         string;
    email:          string;
    userName:       string;
    password:       string;
    activitystatus: string;
    name:           string;
    id?:             number;
  }